backend/requirements.txt
Flask==3.0.3
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-Bcrypt==1.0.1
python-dotenv==1.0.1
requests==2.31.0
shodan==1.31.0
